Move implementation to BeanDefinitionGatherer
