+++ java-frontend/src/test/java/org/sonar/java/model/springcontext/BeanDefinitionGathererTest.java
@@ -26,6 +26,7 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.ArgumentCaptor;
 import org.sonar.api.batch.fs.InputFile;
 import org.sonar.api.batch.sensor.cache.WriteCache;
@@ -36,7 +37,6 @@ import org.sonar.plugins.java.api.ModuleScannerContext;
 import org.sonar.plugins.java.api.caching.CacheContext;
 import org.sonar.plugins.java.api.caching.JavaReadCache;
 import org.sonar.plugins.java.api.caching.JavaWriteCache;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.mockito.ArgumentMatchers.any;
@@ -202,12 +202,10 @@ class BeanDefinitionGathererTest extends SpringContextGathererTest {
     return Stream.of(
       Arguments.of("src/test/files/springcontext/AutowiredDependencies.java", "autowiredDependencies"),
       Arguments.of("src/test/files/springcontext/AutowiredConstructorDependencies.java", "autowiredConstructorDependencies"),
-      Arguments.of("src/test/files/springcontext/BeanMethodWithDependencies.java", "myBean"),
-      Arguments.of("src/test/files/springcontext/SingleConstructorDependencies.java", "singleConstructorDependencies")
+      Arguments.of("src/test/files/springcontext/SingleConstructorDependencies.java", "singleConstructorDependencies"),
+      Arguments.of("src/test/files/springcontext/BeanMethodWithDependencies.java", "myBean")
     );
   }
-
-  // ---- Implicit single-constructor injection --------------------------------
 
   @Test
   void multiple_constructors_without_autowired_yields_no_dependencies() {
@@ -216,25 +214,6 @@ class BeanDefinitionGathererTest extends SpringContextGathererTest {
     var beans = model.getBeanDefinitionRegistry().getByName("multipleConstructorsNoDependencies");
     assertThat(beans).hasSize(1);
     assertThat(beans.get(0).getDependingBeans()).isEmpty();
-  }
-
-  @ParameterizedTest(name = "{0}")
-  @MethodSource("mixedInjectionArguments")
-  void both_injection_sources_collected_when_mixing_injection_styles(String filePath, String beanName) {
-    scan(filePath);
-
-    var beans = model.getBeanDefinitionRegistry().getByName(beanName);
-    assertThat(beans).hasSize(1);
-    var deps = beans.get(0).getDependingBeans();
-    assertThat(deps.get("org.springframework.context.ApplicationContext")).containsOnly("applicationContext");
-    assertThat(deps.get("org.springframework.core.env.Environment")).containsOnly("environment");
-  }
-
-  static Stream<Arguments> mixedInjectionArguments() {
-    return Stream.of(
-      Arguments.of("src/test/files/springcontext/AutowiredConstructorWithUnannotatedConstructor.java", "autowiredConstructorWithUnannotatedConstructor"),
-      Arguments.of("src/test/files/springcontext/MixedInjectionDependencies.java", "mixedInjectionDependencies")
-    );
   }
 
   // ---- @Qualifier handling --------------------------------------------------
@@ -494,6 +473,105 @@ class BeanDefinitionGathererTest extends SpringContextGathererTest {
     var deps = beans.get(0).getDependingBeans();
     assertThat(deps).containsOnlyKeys("org.springframework.context.ApplicationContext");
     assertThat(deps.get("org.springframework.context.ApplicationContext")).containsOnly("applicationContext");
+  }
+
+  // ---- TypeToBeanNamesIndex -------------------------------------------------
+
+  @ParameterizedTest(name = "{0}")
+  @ValueSource(strings = {
+    "src/test/files/springcontext/SimpleComponent.java",
+    "src/test/files/springcontext/SimpleService.java",
+    "src/test/files/springcontext/SimpleRepository.java",
+    "src/test/files/springcontext/SimpleController.java",
+    "src/test/files/springcontext/SimpleRestController.java",
+    "src/test/files/springcontext/SimpleConfiguration.java"
+  })
+  void stereotype_bean_is_registered_under_its_own_type(String filePath) {
+    scan(filePath);
+
+    var index = model.getTypeToBeanNamesIndex();
+    assertThat(index.getNamesForType("checks.spring.context." + beanClassNameFrom(filePath)))
+      .isNotEmpty();
+  }
+
+  @Test
+  void bean_is_registered_under_implemented_interface() {
+    scan("src/test/files/springcontext/ComponentImplementingInterface.java");
+
+    var index = model.getTypeToBeanNamesIndex();
+    assertThat(index.getNamesForType("checks.spring.context.ComponentImplementingInterface"))
+      .containsOnly("componentImplementingInterface");
+    assertThat(index.getNamesForType("org.springframework.context.ApplicationContextAware"))
+      .containsOnly("componentImplementingInterface");
+  }
+
+  @Test
+  void explicit_bean_name_is_used_in_index() {
+    scan("src/test/files/springcontext/ExplicitNameComponent.java");
+
+    var index = model.getTypeToBeanNamesIndex();
+    assertThat(index.getNamesForType("checks.spring.context.ExplicitNameComponent"))
+      .containsOnly("myBean");
+  }
+
+  @Test
+  void bean_method_return_type_is_registered() {
+    scan("src/test/files/springcontext/ConfigurationWithBeanMethods.java");
+
+    var index = model.getTypeToBeanNamesIndex();
+    assertThat(index.getNamesForType("org.springframework.context.ApplicationContext"))
+      .contains("simpleServiceBean", "namedBean", "arrayNamedBean", "emptyNameArrayMethod");
+  }
+
+  @Test
+  void bean_method_aliases_are_all_registered() {
+    scan("src/test/files/springcontext/ConfigurationWithBeanMethods.java");
+
+    // @Bean(name = {"arrayNamedBean", "alias"}) — both names appear in the index
+    var index = model.getTypeToBeanNamesIndex();
+    assertThat(index.getNamesForType("org.springframework.context.ApplicationContext"))
+      .contains("arrayNamedBean", "alias");
+  }
+
+  @Test
+  void multiple_beans_all_registered_in_index() {
+    scan(
+      "src/test/files/springcontext/PayPalProcessor.java",
+      "src/test/files/springcontext/CreditCardProcessor.java"
+    );
+
+    var index = model.getTypeToBeanNamesIndex();
+    assertThat(index.getNamesForType("checks.spring.context.PayPalProcessor"))
+      .containsOnly("paypal");
+    assertThat(index.getNamesForType("checks.spring.context.CreditCardProcessor"))
+      .containsOnly("creditCard");
+  }
+
+  @Test
+  void non_spring_class_registers_nothing_in_index() {
+    scan("src/test/files/springcontext/NoScanAnnotations.java");
+
+    assertThat(model.getTypeToBeanNamesIndex().getNamesForType("checks.spring.context.NoScanAnnotations"))
+      .isEmpty();
+  }
+
+  @Test
+  void index_gatherer_skipped_when_spring_not_in_classpath() {
+    scan(List.of(), "src/test/files/springcontext/SimpleComponent.java");
+
+    assertThat(model.getTypeToBeanNamesIndex().getNamesForType("checks.spring.context.SimpleComponent"))
+      .isEmpty();
+  }
+
+  @Test
+  void index_anonymous_class_is_skipped() {
+    scan("src/test/files/springcontext/SpringBootAppWithAnonymousClass.java");
+
+    assertThat(model.getTypeToBeanNamesIndex().getNamesForType("")).isEmpty();
+  }
+
+  private static String beanClassNameFrom(String filePath) {
+    return filePath.substring(filePath.lastIndexOf('/') + 1, filePath.lastIndexOf('.'));
   }
 
   private static CacheContext mockCacheContext(JavaReadCache readCache, JavaWriteCache writeCache) {
