+++ java-frontend/src/test/java/org/sonar/java/utils/SpringUtilsTest.java
@@ -16,11 +16,7 @@
  */
 package org.sonar.java.utils;
 
-import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.sonar.java.model.JParserTestUtils;
 import org.sonar.java.model.declaration.ClassTreeImpl;
 import org.sonar.java.model.declaration.MethodTreeImpl;
@@ -155,88 +151,4 @@ class SpringUtilsTest {
     assertThat(SpringUtils.isSpringBootUnitTest(method)).isFalse();
   }
 
-  // ---- resolveStereotypeBeanName ----------------------------------------------
-
-  @Test
-  void resolve_stereotype_bean_name_uses_name_attribute() {
-    // Covers the "name".equals(v.name()) branch in the filter
-    var cu = JParserTestUtils.parse("A", """
-      @org.springframework.stereotype.Service(value = "myService")
-      class A {}
-      """, TestClasspathUtils.DEFAULT_MODULE.getClassPath());
-    var clazz = (ClassTreeImpl) cu.types().get(0);
-    assertThat(SpringUtils.resolveStereotypeBeanName(clazz.symbol().metadata(), "A")).isEqualTo("myService");
-  }
-
-  @Test
-  void resolve_stereotype_bean_name_falls_back_to_decapitalized_name() {
-    var cu = JParserTestUtils.parse("A", """
-      @org.springframework.stereotype.Component
-      class MyServiceImpl {}
-      """, TestClasspathUtils.DEFAULT_MODULE.getClassPath());
-    var clazz = (ClassTreeImpl) cu.types().get(0);
-    assertThat(SpringUtils.resolveStereotypeBeanName(clazz.symbol().metadata(), "MyServiceImpl")).isEqualTo("myServiceImpl");
-  }
-
-  // ---- resolveBeanMethodNames -------------------------------------------------
-
-  @ParameterizedTest(name = "{0}")
-  @MethodSource("fallBackToMethodNameArguments")
-  void resolve_bean_method_names_falls_back_to_method_name(String description, String source) {
-    var cu = JParserTestUtils.parse("A", source, TestClasspathUtils.DEFAULT_MODULE.getClassPath());
-    var clazz = (ClassTreeImpl) cu.types().get(0);
-    var method = (MethodTreeImpl) clazz.members().get(0);
-    assertThat(SpringUtils.resolveBeanMethodNames(method)).containsOnly("myMethod");
-  }
-
-  static Stream<Arguments> fallBackToMethodNameArguments() {
-    return Stream.of(
-      Arguments.of("no annotation",           "class A { Object myMethod() { return null; } }"),
-      Arguments.of("empty array",             "import org.springframework.context.annotation.Bean; class A { @Bean(name = {}) Object myMethod() { return null; } }"),
-      Arguments.of("non-name/value attribute","import org.springframework.context.annotation.Bean; class A { @Bean(initMethod = \"init\") Object myMethod() { return null; } }"),
-      Arguments.of("blank name",              "import org.springframework.context.annotation.Bean; class A { @Bean(name = \"\") Object myMethod() { return null; } }")
-    );
-  }
-
-  @Test
-  void resolve_bean_method_names_single_string_name() {
-    var cu = JParserTestUtils.parse("A", """
-      import org.springframework.context.annotation.Bean;
-      class A {
-        @Bean(name = "myBean")
-        Object myMethod() { return null; }
-      }
-      """, TestClasspathUtils.DEFAULT_MODULE.getClassPath());
-    var clazz = (ClassTreeImpl) cu.types().get(0);
-    var method = (MethodTreeImpl) clazz.members().get(0);
-    assertThat(SpringUtils.resolveBeanMethodNames(method)).containsOnly("myBean");
-  }
-
-  @Test
-  void resolve_bean_method_names_array_includes_all_aliases() {
-    var cu = JParserTestUtils.parse("A", """
-      import org.springframework.context.annotation.Bean;
-      class A {
-        @Bean(name = {"primary", "alias"})
-        Object myMethod() { return null; }
-      }
-      """, TestClasspathUtils.DEFAULT_MODULE.getClassPath());
-    var clazz = (ClassTreeImpl) cu.types().get(0);
-    var method = (MethodTreeImpl) clazz.members().get(0);
-    assertThat(SpringUtils.resolveBeanMethodNames(method)).containsExactlyInAnyOrder("primary", "alias");
-  }
-
-  @Test
-  void resolve_bean_method_name_delegates_to_first_name() {
-    var cu = JParserTestUtils.parse("A", """
-      import org.springframework.context.annotation.Bean;
-      class A {
-        @Bean(name = {"first", "second"})
-        Object myMethod() { return null; }
-      }
-      """, TestClasspathUtils.DEFAULT_MODULE.getClassPath());
-    var clazz = (ClassTreeImpl) cu.types().get(0);
-    var method = (MethodTreeImpl) clazz.members().get(0);
-    assertThat(SpringUtils.resolveBeanMethodName(method)).isEqualTo("first");
-  }
 }
